Add job titles
